[#86] Moved history to different endpoint, fixed problem with getting history from backend
