+++ backend/README.md
@@ -24,7 +24,7 @@ Swagger api documentation can be accessed on url `/swagger-ui.html`
 * **Closing session:** /session/close -> session with given ID is now closed (still in DB)
 * **Getting participant list:** /session/participant-list -> returns list of participants (their ID and name)
 * **Getting timeline for one session:** /session/timeline -> returns list of all messages sent in session
-* **Getting whole session history for leader account:** /session/session-history -> returns list of all sessions organized by user
+* **Getting whole session history for leader account:** /session/history -> returns list of all sessions organized by user
 #### Websocket communication in one session:
 * **Creating websocket:** /session-handling (websocket connection address)
 * **Subscribtion:** /topic/session/{sessionId} (receiving all messages, requires getting session ID from connect request)
